feat(storage): 统一存储空间单位为字节并优化上传任务管理
- 将磁盘大小字段从int改为int64，单位统一为字节
- 修改前端磁盘管理界面，实现字节与GB之间的自动转换
- 添加临时文件定时清理任务，清理超过24小时的孤儿临时文件
- 优化上传任务组件，增加防抖机制减少后端接口调用频率
- 实现分片上传重试机制，采用指数退避算法提升稳定性
- 更新用户空间分配逻辑，支持组限制的自动继承
- 修复数据库查询方法，使用First替代Find提升性能

+++ src/core/domain/request/admin_request.go
@@ -114,7 +114,7 @@ type AdminDiskListRequest struct{}
 type AdminCreateDiskRequest struct {
 	DiskPath string `json:"disk_path" binding:"required"`
 	DataPath string `json:"data_path" binding:"required"`
-	Size     int    `json:"size" binding:"required,min=0"` // 大小（GB）
+	Size     int64  `json:"size" binding:"required,min=0"` // 大小（字节）
 }
 
 // AdminUpdateDiskRequest 管理员更新磁盘请求
@@ -122,7 +122,7 @@ type AdminUpdateDiskRequest struct {
 	ID       string `json:"id" binding:"required"`
 	DiskPath string `json:"disk_path"`
 	DataPath string `json:"data_path"`
-	Size     int    `json:"size"`
+	Size     int64  `json:"size"` // 大小（字节）
 }
 
 // AdminDeleteDiskRequest 管理员删除磁盘请求
